+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- Added UIBarButtonItem `actionBlock` property.
+
 ## 3.1.0 (2019-05-23)
 
 - Fixed C++ compiler error